Updates packages that have vulnerability
This patch updates vulnerable packages and enforces following changes

- Modifies few tests and Resources component
- Adds prettier rule for endOfLine as auto to resolve prettier errors
- Moves global variable declaration into a `.d.ts` file
to resolve eslint errors
- Adds `.eslintcache` in the gitignore which is generated by
latest version of react-scripts
