+++ ui/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@patternfly/patternfly": "^4.80.3",
     "@patternfly/react-core": "^4.47.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
@@ -15,20 +16,19 @@
     "@types/react-router-dom": "^5.1.6",
     "@types/react-syntax-highlighter": "^13.5.0",
     "@types/react-test-renderer": "^16.9.3",
-    "axios": "^0.19.2",
+    "axios": "^0.21.1",
     "fuzzysort": "^1.1.4",
     "mobx": "^5.15.6",
     "mobx-react": "^6.3.0",
     "mobx-state-tree": "^3.17.2",
     "moment": "^2.29.1",
     "mst-persist": "^0.1.3",
-    "patternfly": "^3.59.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-github-login": "^1.0.3",
     "react-markdown": "^5.0.3",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "^4.0.1",
     "react-syntax-highlighter": "^15.3.1",
     "remark-gfm": "^1.0.0",
     "typescript": "^3.7.5"
