+++ ui/src/config/constant.d.ts
@@ -1,0 +1,10 @@
+interface API_CONFIG {
+  API_URL: string;
+  GH_CLIENT_ID: string;
+}
+
+export declare global {
+  interface Window {
+    config: API_CONFIG;
+  }
+}
