+++ ui/src/containers/App/App.test.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import renderer from 'react-test-renderer';
 import { FakeHub } from '../../api/testutil';
 import { createProviderAndStore, createProvider } from '../../store/root';
 import LeftPane from '../../components/LeftPane';
@@ -12,13 +11,13 @@ const { Provider } = createProviderAndStore(api);
 
 describe('App', () => {
   it('should render the component correctly and match the snapshot', (done) => {
-    const app = renderer.create(
+    const app = mount(
       <Provider>
         <App />
       </Provider>
     );
 
-    expect(app.toJSON()).toMatchSnapshot();
+    expect(app.debug()).toMatchSnapshot();
     done();
   });
 
@@ -29,7 +28,6 @@ describe('App', () => {
         <App />
       </Provider>
     );
-
     expect(component.find(LeftPane).length).toEqual(1);
   });
 });
